src/Services/IntelliDoc.AI.Service/requirements.txt
fastapi
uvicorn
python-multipart
pytesseract
Pillow
minio
requests
pdf2image
google-generativeai